+++ src/app-toolkit/constants/address.ts
@@ -1,2 +1,3 @@
 export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 export const ETH_ADDR_ALIAS = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee';
+export const DEAD_ADDRESS = '0x000000000000000000000000000000000000dead';