+++ src/apps/angle/ethereum/angle.vault.contract-position-fetcher.ts
@@ -1,17 +1,18 @@
 import { Inject } from '@nestjs/common';
-import { BigNumberish, utils } from 'ethers';
-import { compact, sumBy } from 'lodash';
+import { BigNumberish } from 'ethers';
+import { sum } from 'lodash';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
-import { drillBalance } from '~app-toolkit/helpers/drill-balance.helper';
 import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { DefaultDataProps } from '~position/display.interface';
-import { ContractPositionBalance } from '~position/position-balance.interface';
 import { MetaType } from '~position/position.interface';
-import { isBorrowed, isSupplied } from '~position/position.utils';
-import { GetDisplayPropsParams, GetTokenDefinitionsParams } from '~position/template/contract-position.template.types';
-import { CustomContractPositionTemplatePositionFetcher } from '~position/template/custom-contract-position.template.position-fetcher';
+import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
+import {
+  GetDisplayPropsParams,
+  GetTokenBalancesParams,
+  GetTokenDefinitionsParams,
+} from '~position/template/contract-position.template.types';
 
 import { AngleApiHelper } from '../common/angle.api';
 import { AngleContractFactory, AngleVaultManager } from '../contracts';
@@ -23,7 +24,7 @@ export type AngleVaultDefinition = {
 };
 
 @PositionTemplate()
-export class EthereumAngleVaultsContractPositionFetcher extends CustomContractPositionTemplatePositionFetcher<
+export class EthereumAngleVaultsContractPositionFetcher extends ContractPositionTemplatePositionFetcher<
   AngleVaultManager,
   DefaultDataProps,
   AngleVaultDefinition
@@ -47,56 +48,44 @@ export class EthereumAngleVaultsContractPositionFetcher extends CustomContractPo
     return vaultManagers;
   }
 
-  async getTokenDefinitions({ definition }: GetTokenDefinitionsParams<AngleVaultManager, AngleVaultDefinition>) {
+  async getTokenDefinitions({ contract }: GetTokenDefinitionsParams<AngleVaultManager, AngleVaultDefinition>) {
     return [
-      { metaType: MetaType.SUPPLIED, address: definition.collateral, network: this.network },
-      { metaType: MetaType.BORROWED, address: definition.stablecoin, network: this.network },
+      { metaType: MetaType.SUPPLIED, address: await contract.collateral(), network: this.network },
+      { metaType: MetaType.BORROWED, address: await contract.stablecoin(), network: this.network },
     ];
   }
 
-  async getLabel({
-    contractPosition,
-  }: GetDisplayPropsParams<AngleVaultManager, DefaultDataProps, AngleVaultDefinition>): Promise<string> {
+  async getLabel({ contractPosition }: GetDisplayPropsParams<AngleVaultManager>): Promise<string> {
     return `${getLabelFromToken(contractPosition.tokens[0])} - ${getLabelFromToken(contractPosition.tokens[1])}`;
   }
 
-  async getTokenBalancesPerPosition(): Promise<BigNumberish[]> {
-    throw new Error('Method not implemented.');
-  }
+  async getTokenBalancesPerPosition({
+    address,
+    contract,
+  }: GetTokenBalancesParams<AngleVaultManager>): Promise<BigNumberish[]> {
+    const userBalanceCount = await contract.balanceOf(address);
 
-  async getBalances(address: string): Promise<ContractPositionBalance<DefaultDataProps>[]> {
-    const contractPositions = await this.appToolkit.getAppContractPositions({
-      appId: this.appId,
-      groupIds: [this.groupId],
-      network: this.network,
-    });
+    if (Number(userBalanceCount) < 1) return [0, 0];
 
-    const vaults = Object.values(await this.angleApiHelper.getUserVaults(address, this.network));
+    const vaultIds = Object.values(await this.angleApiHelper.getVaultIds(address, this.network));
 
-    const balances = vaults.map(vault => {
-      const contractPosition = contractPositions.find(v => v.address.toLowerCase() === vault.address.toLowerCase());
-      if (!contractPosition) return null;
+    const balances = await Promise.all(
+      vaultIds.map(async vaultId => {
+        const [collateralData, vaultDebt] = await Promise.all([
+          contract.vaultData(vaultId),
+          contract.getVaultDebt(vaultId),
+        ]);
 
-      const collateralToken = contractPosition!.tokens.find(isSupplied)!;
-      const borrowedToken = contractPosition!.tokens.find(isBorrowed)!;
+        return {
+          supplied: collateralData.collateralAmount,
+          borrowed: vaultDebt,
+        };
+      }),
+    );
 
-      const collateral = utils.parseUnits(vault.collateralAmount.toString(), collateralToken.decimals);
-      const debt = utils.parseUnits(vault.debt.toString(), borrowedToken.decimals);
+    const supplied = sum(balances.map(x => Number(x.supplied)));
+    const borrowed = sum(balances.map(x => Number(x.borrowed)));
 
-      const tokens = [
-        drillBalance(collateralToken, collateral.toString()),
-        drillBalance(borrowedToken, debt.toString(), { isDebt: true }),
-      ];
-
-      const balanceUSD = sumBy(tokens, t => t.balanceUSD);
-
-      return {
-        ...contractPosition,
-        balanceUSD,
-        tokens,
-      };
-    });
-
-    return compact(balances);
+    return [supplied, borrowed];
   }
 }