+++ src/apps/exactly/common/exactly.borrow.token-fetcher.ts
@@ -1,5 +1,3 @@
-import type { IMulticallWrapper } from '~multicall/multicall.interface';
-import type { AppTokenPosition } from '~position/position.interface';
 import type { GetDataPropsParams, GetTokenPropsParams } from '~position/template/app-token.template.types';
 
 import type { ExactlyMarketDefinition } from '../common/exactly.definitions-resolver';
@@ -20,23 +18,5 @@ export abstract class ExactlyBorrowFetcher extends ExactlyTokenFetcher {
 
   getApr({ definition }: GetDataPropsParams<Market, ExactlyMarketProps, ExactlyMarketDefinition>) {
     return Number(definition.floatingBorrowRate) / 1e16;
-  }
-
-  async getBalancePerToken({
-    address,
-    appToken,
-    multicall,
-  }: {
-    address: string;
-    appToken: AppTokenPosition;
-    multicall: IMulticallWrapper;
-  }) {
-    const { floatingBorrowShares } = await this.definitionsResolver.getDefinition({
-      multicall,
-      network: this.network,
-      account: address,
-      market: appToken.address,
-    });
-    return floatingBorrowShares;
   }
 }