+++ src/apps/granary-finance/optimism/granary-finance.position-presenter.ts
@@ -1,7 +1,29 @@
 import { PresenterTemplate } from '~app-toolkit/decorators/presenter-template.decorator';
+import { PresentationConfig } from '~app/app.interface';
 import { AaveV2PositionPresenter } from '~apps/aave-v2/common/aave-v2.position-presenter';
 
 @PresenterTemplate()
 export class OptimismGranaryFinancePositionPresenter extends AaveV2PositionPresenter {
   lendingPoolAddress = '0x8fd4af47e4e63d1d2d45582c3286b4bd9bb95dfe';
+
+  explorePresentationConfig?: PresentationConfig = {
+    tabs: [
+      {
+        label: 'Lending',
+        viewType: 'split',
+        views: [
+          {
+            viewType: 'list',
+            label: 'Supply',
+            groupIds: ['supply'],
+          },
+          {
+            viewType: 'list',
+            label: 'Borrow',
+            groupIds: ['variable-debt'],
+          },
+        ],
+      },
+    ],
+  };
 }